add MacCheckpoint class
